+++ pdfcrowd.php
@@ -387,7 +387,7 @@ Possible reasons:
 
     private $fields, $scheme, $port, $api_prefix, $curlopt_timeout;
 
-    public static $client_version = "6.1.0";
+    public static $client_version = "6.2.0";
     public static $http_port = 80;
     public static $https_port = 443;
     public static $api_host = 'pdfcrowd.com';
@@ -547,7 +547,7 @@ You need to restart your web server after installation.';
         $this->reset_response_data();
         $this->setProxy(null, null, null, null);
         $this->setUseHttp(false);
-        $this->setUserAgent('pdfcrowd_php_client/6.1.0 (https://pdfcrowd.com)');
+        $this->setUserAgent('pdfcrowd_php_client/6.2.0 (https://pdfcrowd.com)');
 
         $this->retry_count = 1;
         $this->converter_version = '24.04';
@@ -595,7 +595,7 @@ You need to restart your web server after installation.';
 
     private static $SSL_ERRORS = array(35, 51, 53, 54, 58, 59, 60, 64, 66, 77, 80, 82, 83, 90, 91);
 
-    const CLIENT_VERSION = '6.1.0';
+    const CLIENT_VERSION = '6.2.0';
     public static $MULTIPART_BOUNDARY = '----------ThIs_Is_tHe_bOUnDary_$';
 
     private function add_file_field($name, $file_name, $data, &$body) {
@@ -1329,12 +1329,12 @@ class HtmlToPdfClient {
     /**
     * Set the page range to print.
     *
-    * @param pages A comma separated list of page numbers or ranges.
+    * @param pages A comma separated list of page numbers or ranges. Special strings may be used, such as `odd`, `even` and `last`.
     * @return The converter object.
     */
     function setPrintPageRange($pages) {
-        if (!preg_match("/^(?:\s*(?:\d+|(?:\d*\s*\-\s*\d+)|(?:\d+\s*\-\s*\d*))\s*,\s*)*\s*(?:\d+|(?:\d*\s*\-\s*\d+)|(?:\d+\s*\-\s*\d*))\s*$/", $pages))
-            throw new Error(create_invalid_value_message($pages, "setPrintPageRange", "html-to-pdf", "A comma separated list of page numbers or ranges.", "set_print_page_range"), 470);
+        if (!preg_match("/^(?:\s*(?:\d+|(?:\d*\s*\-\s*\d+)|(?:\d+\s*\-\s*\d*)|odd|even|last)\s*,\s*)*\s*(?:\d+|(?:\d*\s*\-\s*\d+)|(?:\d+\s*\-\s*\d*)|odd|even|last)\s*$/", $pages))
+            throw new Error(create_invalid_value_message($pages, "setPrintPageRange", "html-to-pdf", "A comma separated list of page numbers or ranges. Special strings may be used, such as `odd`, `even` and `last`.", "set_print_page_range"), 470);
         
         $this->fields['print_page_range'] = $pages;
         return $this;
@@ -1514,7 +1514,7 @@ class HtmlToPdfClient {
     }
 
     /**
-    * The page header is not printed on the specified pages.
+    * The page header content is not printed on the specified pages. To remove the entire header area, use the <a href="#set_conversion_config">conversion config</a>.
     *
     * @param pages List of physical page numbers. Negative numbers count backwards from the last page: -1 is the last page, -2 is the last but one page, and so on. A comma separated list of page numbers.
     * @return The converter object.
@@ -1528,7 +1528,7 @@ class HtmlToPdfClient {
     }
 
     /**
-    * The page footer is not printed on the specified pages.
+    * The page footer content is not printed on the specified pages. To remove the entire footer area, use the <a href="#set_conversion_config">conversion config</a>.
     *
     * @param pages List of physical page numbers. Negative numbers count backwards from the last page: -1 is the last page, -2 is the last but one page, and so on. A comma separated list of page numbers.
     * @return The converter object.
@@ -2888,12 +2888,15 @@ The structure of the JSON must be:
     <ul>
       <li>
       <em>pages</em>:
-        A comma-separated list of page numbers or ranges. For example:
+        A comma-separated list of page numbers or ranges.
+        Special strings may be used, such as `odd`, `even` and `last`.
+        For example:
       <ul>
       <li><em>1-</em>: from page 1 to the end of the document</li>
       <li><em>2</em>: only the 2nd page</li>
-      <li><em>2, 4, 6</em>: pages 2, 4, and 6</li>
+      <li><em>2,4,6</em>: pages 2, 4, and 6</li>
       <li><em>2-5</em>: pages 2 through 5</li>
+      <li><em>odd,2</em>: the 2nd page and all odd pages</li>
       </ul>
       </li>
       <li><em>pageSize</em>: The page size (optional).
@@ -6804,6 +6807,20 @@ class PdfToHtmlClient {
     }
 
     /**
+    * Sets the processing mode for handling Type 3 fonts.
+    *
+    * @param mode The type3 font mode. Allowed values are raster, convert.
+    * @return The converter object.
+    */
+    function setType3Mode($mode) {
+        if (!preg_match("/(?i)^(raster|convert)$/", $mode))
+            throw new Error(create_invalid_value_message($mode, "setType3Mode", "pdf-to-html", "Allowed values are raster, convert.", "set_type3_mode"), 470);
+        
+        $this->fields['type3_mode'] = $mode;
+        return $this;
+    }
+
+    /**
     * Converts ligatures, two or more letters combined into a single glyph, back into their individual ASCII characters.
     *
     * @param value Set to <span class='field-value'>true</span> to split ligatures.
@@ -6811,6 +6828,20 @@ class PdfToHtmlClient {
     */
     function setSplitLigatures($value) {
         $this->fields['split_ligatures'] = $value;
+        return $this;
+    }
+
+    /**
+    * Apply custom CSS to the output HTML document. It allows you to modify the visual appearance and layout. Tip: Using <span class='field-value'>!important</span> in custom CSS provides a way to prioritize and override conflicting styles.
+    *
+    * @param css A string containing valid CSS. The string must not be empty.
+    * @return The converter object.
+    */
+    function setCustomCss($css) {
+        if (!($css != null && $css !== ''))
+            throw new Error(create_invalid_value_message($css, "setCustomCss", "pdf-to-html", "The string must not be empty.", "set_custom_css"), 470);
+        
+        $this->fields['custom_css'] = $css;
         return $this;
     }